feedback: markus
* move cachedir() back to CLI
* make cache dir an input for sdk
* rename GetManifestsResponse to CoordinatorState
* rename hostData to coordinatorPolicyChecksum
